+++ shared/core/designSystem/src/androidMain/AndroidManifest.xml
