+++ shared/core/designSystem/src/androidMain/kotlin/com/kus/designsystem/Platform.android.kt
@@ -1,0 +1,3 @@
+package com.kus.designsystem
+
+actual fun platform() = "Android"
