+++ shared/core/designSystem/src/commonMain/kotlin/com/kus/designsystem/Platform.kt
@@ -1,0 +1,3 @@
+package com.kus.designsystem
+
+expect fun platform(): String
