+++ shared/core/designSystem/src/desktopMain/kotlin/com/kus/designsystem/Platform.desktop.kt
@@ -1,0 +1,3 @@
+package com.kus.designsystem
+
+actual fun platform(): String = "Desktop"
