+++ shared/core/designSystem/src/iosMain/kotlin/com/kus/designsystem/Platform.ios.kt
@@ -1,0 +1,3 @@
+package com.kus.designsystem
+
+actual fun platform() = "iOS"
