--- shared/core/src/androidHostTest/kotlin/com/kus/core/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.kus.core
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
