--- shared/core/src/androidMain/kotlin/com/kus/core/Platform.android.kt
@@ -1,3 +1,0 @@
-package com.kus.core
-
-actual fun platform() = "Android"
