--- shared/core/src/commonMain/kotlin/com/kus/core/Platform.kt
@@ -1,3 +1,0 @@
-package com.kus.core
-
-expect fun platform(): String
