--- shared/core/src/desktopMain/kotlin/com/kus/core/Platform.desktop.kt
@@ -1,3 +1,0 @@
-package com.kus.core
-
-actual fun platform(): String = "Desktop"
