--- shared/core/src/iosMain/kotlin/com/kus/core/Platform.ios.kt
@@ -1,3 +1,0 @@
-package com.kus.core
-
-actual fun platform() = "iOS"
